+++ clients/sub_account/src/binance_sdk_sub_account/rest_api/models/move_position_for_sub_account_response_move_position_orders_inner.py
@@ -33,7 +33,9 @@ class MovePositionForSubAccountResponseMovePositionOrdersInner(BaseModel):
     price_type: Optional[StrictStr] = Field(default=None, alias="priceType")
     price: Optional[StrictStr] = None
     quantity: Optional[StrictStr] = None
-    position_side: Optional[StrictStr] = Field(default=None, alias="positionSide")
+    position_side: Optional[StrictStr] = Field(
+        default=None, description="null when productType=OPTION.", alias="positionSide"
+    )
     side: Optional[StrictStr] = None
     success: Optional[StrictBool] = None
     additional_properties: Dict[str, Any] = {}
@@ -100,6 +102,11 @@ class MovePositionForSubAccountResponseMovePositionOrdersInner(BaseModel):
         if self.additional_properties is not None:
             for _key, _value in self.additional_properties.items():
                 _dict[_key] = _value
+
+        # set to None if position_side (nullable) is None
+        # and model_fields_set contains the field
+        if self.position_side is None and "position_side" in self.model_fields_set:
+            _dict["positionSide"] = None
 
         return _dict